Updating truth kinematics file parsing and plotting scripts to reflect general repo restructure.
