+++ truth_kinematics/file_parsing/beam_background_kinematics.py
@@ -1,15 +1,13 @@
-import matplotlib
-import matplotlib.pyplot as plt
 import h5py
 import argparse
 import numpy as np
-import twoBytwo_defs
-import mip_backgrounds
-import auxiliary
 import glob
-
-nu_signal_pdg=-14
-pion_pdg={111,211,-211}
+import mip_backgrounds
+import sys
+sys.path.append('../../common')
+import file_parsing
+import geometry_methods as geo_methods
+import truth_methods as truth
 
 
 def main(sim_dir, input_type):
@@ -30,22 +28,22 @@ def main(sim_dir, input_type):
         ### partition file by spill
         unique_spill = np.unique(sim_h5['trajectories']['spillID'])
         for spill_id in unique_spill:
-            ghdr, gstack, traj, vert, seg = auxiliary.get_spill_data(sim_h5, spill_id, input_type)
+            ghdr, gstack, traj, vert, seg = file_parsing.get_spill_data(sim_h5, spill_id, input_type)
 
             ### partition by vertex ID within beam spill
             for v_i in range(len(vert['vertexID'])):
                 vert_pos= [vert['x_vert'][v_i], vert['y_vert'][v_i], vert['z_vert'][v_i]]
-                vert_in_active_LAr = twoBytwo_defs.fiducialized_vertex( vert_pos )
+                vert_in_active_LAr = geo_methods.fiducialized_vertex( vert_pos )
 
                 ##### REQUIRE neutrino vertex in LAr active volume #####
                 if vert_in_active_LAr==False: continue
 
                 vert_id = vert['vertexID'][v_i]
             
-                nu_mu_bar = auxiliary.signal_nu_pdg(ghdr, vert_id)
-                is_cc = auxiliary.signal_cc(ghdr, vert_id)
-                pionless = auxiliary.signal_meson_status(gstack, vert_id)
-                fv_particle_origin=twoBytwo_defs.fiducialized_particle_origin(traj, vert_id)
+                nu_mu_bar = truth.signal_nu_pdg(ghdr, vert_id)
+                is_cc = truth.signal_cc(ghdr, vert_id)
+                pionless = truth.signal_meson_status(gstack, vert_id)
+                fv_particle_origin=geo_methods.fiducialized_particle_origin(traj, vert_id)
                         
                 ##### THRESHOLD BACKGROUNDS #####
                 if is_cc==True and pionless==False and fv_particle_origin==True:
@@ -57,8 +55,8 @@ def main(sim_dir, input_type):
 #                    mip_backgrounds.pion_characterization(spill_id, vert_id, ghdr, gstack, traj, vert, seg, nc_dict)
 #                    mip_backgrouns.primaries(spill_id, vert_id, ghdr, gstack, traj, vert, seg, nc_primaries_dict)
 
-    auxiliary.save_dict_to_json(cc_dict, 'cc_pion_backgrounds', True)
-    auxiliary.save_dict_to_json(cc_primaries_dict, 'cc_primaries', True)
+    file_parsing.save_dict_to_json(cc_dict, 'cc_pion_backgrounds', True)
+    file_parsing.save_dict_to_json(cc_primaries_dict, 'cc_primaries', True)
 #    auxiliary.save_dict_to_json(nc_dict, 'nc_pion_backgrounds', True)
 #    auxiliary.save_dict_to_json(nc_primaries_dict, 'nc_primaries', True)
