+++ truth_kinematics/file_parsing/cc_threshold_bkg_dict.json
