+++ truth_kinematics/file_parsing/dirt_bkg_dict.json
