+++ truth_kinematics/file_parsing/signal_dict.json
