+++ truth_kinematics/plotting/plot_stacked_kinematics.py
